# manager-external-client/api/openapi.yaml
openapi: 3.0.0
info:
  description: Public API for the Preservation Manager.
  title: Preservation Manager Public API
  version: 1.0.0
servers:
- description: Local server
  url: http://localhost:8484/api
security:
- ApiKeyAuth: []
...
paths:
...
  /batch/{ingestId}:
    get:
      operationId: retrieveBatch
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: path
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveBatchResponse'
          description: Ingest Batch details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves Ingest Batch details
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/{ingestId}/objects:
    get:
      operationId: retrieveBatchObjects
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: path
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The number of results to include per page (default 20)
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
      - description: The result page to retrieve, 0 indexed
        example: 0
        in: query
        name: page
        required: false
        schema:
          type: integer
      - description: Filter objects by the existence of errors or warnings. WARNINGS
          returns objects that only have warnings; ERRORS returns any object that
          has errors.
        example: ERRORS
        in: query
        name: hasProblems
        required: false
        schema:
          enum:
          - NONE
          - WARNINGS
          - ERRORS
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveBatchObjectsResponse'
          description: Ingest Batch Object details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves Ingest Batch Objects
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/object:
    get:
      operationId: retrieveBatchObject
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: query
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveBatchObjectResponse'
          description: Ingest Batch Object details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves details about an Ingest Batch Object
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/object/files:
    get:
      operationId: retrieveBatchObjectFiles
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: query
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveBatchObjectFilesResponse'
          description: Ingest Batch Object files
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves the files associated to an Ingest Batch Object
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/object/file:
    get:
      operationId: retrieveBatchObjectFile
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: query
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      - description: The object relative path of the file
        example: path/to/file.txt
        in: query
        name: filePath
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveBatchObjectFileResponse'
          description: Ingest Batch Object file
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves a specific file associated to an Ingest Batch Object
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/object/diff:
    get:
      operationId: diffBatchObject
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: query
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffBatchObjectResponse'
          description: Batch object diff
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Diffs the batch object against the current state of the object in the
        preservation system
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/reject:
    post:
      operationId: rejectIngestBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectIngestBatchRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Rejects a batch for ingestion
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/approve:
    post:
      operationId: approveIngestBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveIngestBatchRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Approves a batch for ingestion
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/retryIngest:
    post:
      operationId: retryBatchIngest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRetryIngestRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retries a batch ingest if it has failed
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/retryReplicate:
    post:
      operationId: retryBatchReplicate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRetryReplicateRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retries a batch replicate if it has failed
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/object/approve:
    post:
      operationId: approveIngestObject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveIngestObjectRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Approves an object for ingestion
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/object/reject:
    post:
      operationId: rejectIngestObject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectIngestObjectRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Rejects an object for ingestion
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /batch/{ingestId}/event:
    get:
      operationId: retrieveBatchEvents
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: path
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveEventsResponse'
          description: Batch events
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves all of the events associated to the batch. This does not
        include events associated to objects in the batch
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/object/event:
    get:
      operationId: retrieveBatchObjectEvents
      parameters:
      - description: The ID for the Ingest Batch
        example: 10293
        in: query
        name: ingestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveEventsResponse'
          description: Object events
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves all of the events associated to the object in the batch
      tags:
      - PreservationManager
      x-accepts: application/json
  /batch/search:
    get:
      operationId: searchBatches
      parameters:
      - description: The ID of the organization
        example: organization-id
        in: query
        name: orgName
        required: false
        schema:
          type: string
      - description: The vault an object is in
        example: vault-name
        explode: false
        in: query
        name: vault
        required: false
        schema:
          items:
            type: string
          type: array
      - description: The state the batch is in
        example: PENDING_REVIEW
        explode: false
        in: query
        name: state
        required: false
        schema:
          items:
            enum:
            - RECEIVED
            - ANALYZING
            - ANALYSIS_FAILED
            - PENDING_REVIEW
            - PENDING_INGESTION
            - PENDING_REJECTION
            - INGESTING
            - INGEST_FAILED
            - REPLICATING
            - REPLICATION_FAILED
            - COMPLETE
            - REJECTING
            - REJECTED
            - DELETED
            type: string
          type: array
      - description: The number of results to include per page (default 20)
        example: 20
        in: query
        name: pageSize
        required: false
        schema:
          type: integer
      - description: The result page to retrieve, 0 indexed
        example: 0
        in: query
        name: page
        required: false
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSearchResponse'
          description: Batch search results sorted by the batch updated timestamp
            in reverse chronological order
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Search for batches
      tags:
      - PreservationManager
      x-accepts: application/json
...
  /object:
    get:
      operationId: describePreservationObject
      parameters:
      - description: The vault an object is in
        example: vault-name
        in: query
        name: vault
        required: true
        schema:
          type: string
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      - description: The version of an object
        example: 1
        in: query
        name: version
        required: false
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribePreservationObjectResponse'
          description: Details about a preservation object
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves details about an object version
      tags:
      - PreservationManager
      x-accepts: application/json
  /object/event:
    get:
      operationId: retrieveObjectEvents
      parameters:
      - description: The vault an object is in
        example: vault-name
        in: query
        name: vault
        required: true
        schema:
          type: string
      - description: The external ID of the object
        example: 1711.dl/ZQ5WPGU2GKBFT8Q
        in: query
        name: externalObjectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveEventsResponse'
          description: Object event logs
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves all of the events associated with an object
      tags:
      - PreservationManager
      x-accepts: application/json
...
components:
  schemas:
...
    RetrieveBatchResponse:
      example:
        ingestBatch:
          orgName: organization-id
          hasAnalysisWarnings: true
          receivedTimestamp: 2018-12-27T20:50:06.907193Z
          createdBy: createdBy
          fileSize: 1024
          hasAnalysisErrors: true
          reviewedBy: reviewedBy
          updatedTimestamp: 2018-12-27T20:50:09.189558Z
          ingestId: 10293
          vault: vault-name
          originalFilename: bag.zip
      properties:
        ingestBatch:
          $ref: '#/components/schemas/IngestBatchSummary'
      required:
      - ingestBatch
...
    RetrieveBatchObjectResponse:
      example:
        batchObject:
          hasAnalysisWarnings: true
          hasAnalysisErrors: true
          externalObjectId: 1711.dl/ZQ5WPGU2GKBFT8Q
          reviewedBy: reviewedBy
          version: 1
          ingestId: 10293
          vault: vault-name
      properties:
        batchObject:
          $ref: '#/components/schemas/IngestBatchObject'
      required:
      - batchObject
...
    RejectIngestBatchRequest:
      example:
        ingestId: 10293
      properties:
        ingestId:
          description: The ID for the Ingest Batch
          example: 10293
          format: int64
          type: integer
      required:
      - ingestId
...
    ErrorResponse:
      properties:
        timestamp:
          description: Timestamp when the error occurred
          format: date-time
          type: string
        requestId:
          description: UUID assigned to the request
          example: 6037de0a4b2e4611956da540938a00ff
          type: string
        status:
          description: HTTP status code
          example: 500
          format: int32
          type: integer
        error:
          description: Application classification of the error
          type: string
        message:
          description: Description of the error
          example: Invalid input
          type: string
        path:
          description: Request path
          example: /api/ingest/bag
          type: string
      required:
      - error
      - message
      - path
      - requestId
      - status
      - timestamp
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-KEY
      type: apiKey